+++ packages/app/playwright/crx/utils/popup.ts
@@ -26,14 +26,17 @@ export async function switchAccount(popupPage: Page, name: string) {
   }
 
   await getByAriaLabel(popupPage, 'Accounts').click();
-  // Add position to click on the element and not on copy button
+
+  await popupPage.waitForTimeout(5000);
   await hasText(popupPage, name);
+  // Add position to click on the element and not on copy button
   await getByAriaLabel(popupPage, name).click({
     position: {
       x: 10,
       y: 10,
     },
   });
+  await popupPage.waitForTimeout(2000);
   await waitAriaLabel(popupPage, `${name} selected`);
 
   // Return account to be used on tests