+++ packages/app/src/systems/Asset/components/AssetItem/AssetItem.tsx
@@ -1,6 +1,7 @@
 import { cssObj } from '@fuel-ui/css';
 import {
   Avatar,
+  Badge,
   Box,
   Button,
   CardList,
@@ -35,6 +36,7 @@ export type AssetItemProps = {
   showActions?: boolean;
   onRemove?: (assetId: string) => void;
   onEdit?: (assetId: string) => void;
+  shouldShowAddAssetBtn?: boolean;
 };
 
 type AssetItemComponent = FC<AssetItemProps> & {
@@ -48,6 +50,7 @@ export const AssetItem: AssetItemComponent = ({
   showActions,
   onRemove,
   onEdit,
+  shouldShowAddAssetBtn,
 }) => {
   const navigate = useNavigate();
   const { visibility } = useBalanceVisibility();
@@ -126,7 +129,9 @@ export const AssetItem: AssetItemComponent = ({
           delayDuration={0}
           open={visibility && tooltip ? undefined : false}
         >
-          <Text css={{ fontSize: '$sm', fontWeight: '$normal' }}>
+          <Text
+            css={{ fontSize: '$sm', fontWeight: '$normal', textAlign: 'right' }}
+          >
             <AmountVisibility
               value={amount}
               units={decimals}
@@ -141,9 +146,22 @@ export const AssetItem: AssetItemComponent = ({
     return null;
   }
 
-  function goToAsset() {
-    navigate(Pages.assetsAdd({ assetId }));
+  function goToAsset(asset: {
+    assetId?: string;
+    name?: string;
+    symbol?: string;
+    decimals?: number;
+  }) {
+    navigate(Pages.assetsAdd({ assetId }), { state: asset });
   }
+
+  const suspiciousTooltipContent = (
+    <div style={{ textAlign: 'center' }}>
+      This asset is flagged as suspicious,
+      <br /> it may mimicking another asset.
+      <br /> Proceed with caution.
+    </div>
+  );
 
   return (
     <CardList.Item rightEl={getRightEl()} css={{ alignItems: 'center' }}>
@@ -161,20 +179,35 @@ export const AssetItem: AssetItemComponent = ({
       )}
       <Box.Flex direction="column">
         <Heading as="h6" css={styles.assetName}>
-          {name || (
-            <Box.Flex>
-              Unknown
+          <Box.Flex>
+            {name || 'Unknown'}
+            {asset.suspicious ? (
+              <Tooltip content={suspiciousTooltipContent}>
+                <Icon
+                  css={styles.assetSuspicious}
+                  icon={Icon.is('AlertTriangle')}
+                />
+              </Tooltip>
+            ) : (
+              ''
+            )}
+            {asset.isNft && (
+              <Badge variant="ghost" intent="primary" css={styles.assetNft}>
+                NFT
+              </Badge>
+            )}
+            {shouldShowAddAssetBtn && (
               <Button
                 size="xs"
                 intent="primary"
                 variant="link"
-                onPress={goToAsset}
+                onPress={() => goToAsset(asset)}
                 css={styles.addAssetBtn}
               >
                 (Add)
               </Button>
-            </Box.Flex>
-          )}
+            )}
+          </Box.Flex>
         </Heading>
         <Box.Flex direction="row">
           {symbol ? (
@@ -207,6 +240,11 @@ const styles = {
     textSize: 'sm',
     fontWeight: '$normal',
   }),
+  assetSuspicious: cssObj({
+    marginLeft: 5,
+    marginRight: 5,
+    color: 'orange',
+  }),
   addAssetBtn: cssObj({
     p: '0',
     ml: '$1',
@@ -228,5 +266,10 @@ const styles = {
     '.fuel_Button:hover': {
       color: '$intentsBase11 !important',
     },
+  }),
+  assetNft: cssObj({
+    ml: '$2',
+    fontSize: '$sm',
+    lineHeight: 'normal',
   }),
 };