+++ packages/app/src/systems/Core/utils/databaseVersioning.ts
@@ -187,6 +187,7 @@ export const applyDbVersioning = (db: Dexie) => {
     });
 
   // DB VERSION 26
+  // cleanup networks and add defaults
   db.version(26)
     .stores({
       vaults: 'key',
@@ -217,4 +218,18 @@ export const applyDbVersioning = (db: Dexie) => {
         });
       }
     });
+
+  // DB VERSION 27
+  // add indexedAssets table
+  db.version(27).stores({
+    vaults: 'key',
+    accounts: '&address, &name',
+    networks: '&id, &url, &name, chainId',
+    connections: 'origin',
+    transactionsCursors: '++id, address, size, providerUrl, endCursor',
+    assets: '&name, &symbol',
+    indexedAssets: 'key',
+    abis: '&contractId',
+    errors: '&id',
+  });
 };