+++ packages/e2e-contract-tests/playwright/e2e/utils/contract.ts
@@ -1,5 +1,10 @@
 import type { FuelWalletTestHelper } from '@fuels/playwright-utils';
-import { getButtonByText, getByAriaLabel } from '@fuels/playwright-utils';
+import {
+  expect,
+  getButtonByText,
+  getByAriaLabel,
+  hasText,
+} from '@fuels/playwright-utils';
 import type { Page } from '@playwright/test';
 
 export const connect = async (
@@ -11,4 +16,15 @@ export const connect = async (
   await connectButton.click();
   await getByAriaLabel(page, `Connect to ${walletName}`, true).click();
   await fuelWalletTestHelper.walletConnect();
+
+  await expect
+    .poll(
+      () =>
+        hasText(page, 'Status: Connected')
+          .then(() => true)
+          .catch(() => false),
+      { timeout: 15000 }
+    )
+    .toBeTruthy();
+  await page.waitForTimeout(3000);
 };