+++ packages/e2e-contract-tests/src/contracts/contracts/index.ts
@@ -5,9 +5,9 @@
 /* eslint-disable @typescript-eslint/consistent-type-imports */
 
 /*
-  Fuels version: 0.95.0
+  Fuels version: 0.96.1
   Forc version: 0.65.2
-  Fuel-Core version: 0.37.1
+  Fuel-Core version: 0.38.0
 */
 
 export { CustomAsset } from './CustomAsset';