+++ packages/types/src/asset.ts
@@ -4,6 +4,10 @@ export type AssetData = Asset & {
   // override icon to don't be required
   icon?: string;
   isCustom?: boolean;
+  indexed?: boolean;
+  suspicious?: boolean;
+  isNft?: boolean;
+  verified?: boolean;
 };
 
 export type AssetAmount = AssetData & {
@@ -14,6 +18,10 @@ export type AssetFuelData = AssetFuel & {
   // override icon to don't be required
   icon?: string;
   isCustom?: boolean;
+  indexed?: boolean;
+  suspicious?: boolean;
+  isNft?: boolean;
+  verified?: boolean;
 };
 
 export type AssetFuelAmount = AssetFuelData & {