+++ packages/types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @fuel-wallet/types
 
+## 0.41.1
+
+### Patch Changes
+
+- [#1648](https://github.com/FuelLabs/fuels-wallet/pull/1648) [`e63d6eba`](https://github.com/FuelLabs/fuels-wallet/commit/e63d6ebaecf68340e2b0cd690056cea2bd317b29) Thanks [@LuizAsFight](https://github.com/LuizAsFight)! - feat: improve nft support + dont allow sending funds to assetId address
+
+## 0.41.0
+
+### Minor Changes
+
+- [#1629](https://github.com/FuelLabs/fuels-wallet/pull/1629) [`8a20dded`](https://github.com/FuelLabs/fuels-wallet/commit/8a20ddedbebe2c49c985d414766f2307e275db8b) Thanks [@rodrigobranas](https://github.com/rodrigobranas)! - added SRC20 custom assets name, symbol and decimal resolve from indexer
+
 ## 0.40.1
 
 ## 0.40.0